requirements.txt
django==5.2
djangorestframework==3.16.0
requests==2.31.0